backend/lambda/auth/requirements.txt
bcrypt==4.0.1
PyJWT==2.8.0
boto3==1.26.137
botocore==1.29.137
